Upgrade Prow infrastructure
Add the Prow cluster manifests.

We are going to upgrade our Prow infra to latest stable. This adds
the `crier` component which is doing everything that the our
`bootstrap.py` file was doing. However, it's preferable to use
native Prow tools for the testing infrastructure.

Also, refactored the existing Prow config, and separate the general
config, and the job configs. Now, we have a `Makefile` to easily update
the Prow configs.