Remove `--flake-attempts` parameter
Kubernetes 1.25 or newer E2E tests use Ginkgo v2. This version
doesn't properly include the flake tests in the junit report.

Therefore, the testgrid dashboard loses the visibility of the
flake tests. It will render the flake tests with green color, similar
with the tests that didn't flake at all.

Signed-off-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>