Use a script to statically update the `.vscode/launch.json` file
Previously, we used a hack that relied on the Docker VSCode plugin to
call an external script when the args are passed to the `docker run`
command.

While that worked on Linux/MacOS, it doesn't work on Windows.

This commits makes the `vscode/launch.json` a static file that can be
updated fia `prow/update-vscode-debug-targets.py` script.

Signed-off-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>